+++ apps/greenhouse/src/routeTree.gen.ts
@@ -8,27 +8,46 @@
 // You should NOT make any changes in this file as it will be overwritten.
 // Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
 
-// Import Routes
+import { Route as rootRouteImport } from "./routes/__root"
+import { Route as IndexRouteImport } from "./routes/index"
+import { Route as ExtensionIdSplatRouteImport } from "./routes/$extensionId.$"
 
-import { Route as rootRoute } from "./routes/__root"
-import { Route as IndexImport } from "./routes/index"
-import { Route as ExtensionIdImport } from "./routes/$extensionId.*"
-
-// Create/Update Routes
-
-const IndexRoute = IndexImport.update({
+const IndexRoute = IndexRouteImport.update({
   id: "/",
   path: "/",
-  getParentRoute: () => rootRoute,
+  getParentRoute: () => rootRouteImport,
+} as any)
+const ExtensionIdSplatRoute = ExtensionIdSplatRouteImport.update({
+  id: "/$extensionId/$",
+  path: "/$extensionId/$",
+  getParentRoute: () => rootRouteImport,
 } as any)
 
-const ExtensionIdRoute = ExtensionIdImport.update({
-  id: "/$extensionId/*",
-  path: "/$extensionId/*",
-  getParentRoute: () => rootRoute,
-} as any)
-
-// Populate the FileRoutesByPath interface
+export interface FileRoutesByFullPath {
+  "/": typeof IndexRoute
+  "/$extensionId/$": typeof ExtensionIdSplatRoute
+}
+export interface FileRoutesByTo {
+  "/": typeof IndexRoute
+  "/$extensionId/$": typeof ExtensionIdSplatRoute
+}
+export interface FileRoutesById {
+  __root__: typeof rootRouteImport
+  "/": typeof IndexRoute
+  "/$extensionId/$": typeof ExtensionIdSplatRoute
+}
+export interface FileRouteTypes {
+  fileRoutesByFullPath: FileRoutesByFullPath
+  fullPaths: "/" | "/$extensionId/$"
+  fileRoutesByTo: FileRoutesByTo
+  to: "/" | "/$extensionId/$"
+  id: "__root__" | "/" | "/$extensionId/$"
+  fileRoutesById: FileRoutesById
+}
+export interface RootRouteChildren {
+  IndexRoute: typeof IndexRoute
+  ExtensionIdSplatRoute: typeof ExtensionIdSplatRoute
+}
 
 declare module "@tanstack/react-router" {
   interface FileRoutesByPath {
@@ -36,74 +55,21 @@ declare module "@tanstack/react-router" {
       id: "/"
       path: "/"
       fullPath: "/"
-      preLoaderRoute: typeof IndexImport
-      parentRoute: typeof rootRoute
+      preLoaderRoute: typeof IndexRouteImport
+      parentRoute: typeof rootRouteImport
     }
-    "/$extensionId/*": {
-      id: "/$extensionId/*"
-      path: "/$extensionId/*"
-      fullPath: "/$extensionId/*"
-      preLoaderRoute: typeof ExtensionIdImport
-      parentRoute: typeof rootRoute
+    "/$extensionId/$": {
+      id: "/$extensionId/$"
+      path: "/$extensionId/$"
+      fullPath: "/$extensionId/$"
+      preLoaderRoute: typeof ExtensionIdSplatRouteImport
+      parentRoute: typeof rootRouteImport
     }
   }
-}
-
-// Create and export the route tree
-
-export interface FileRoutesByFullPath {
-  "/": typeof IndexRoute
-  "/$extensionId/*": typeof ExtensionIdRoute
-}
-
-export interface FileRoutesByTo {
-  "/": typeof IndexRoute
-  "/$extensionId/*": typeof ExtensionIdRoute
-}
-
-export interface FileRoutesById {
-  __root__: typeof rootRoute
-  "/": typeof IndexRoute
-  "/$extensionId/*": typeof ExtensionIdRoute
-}
-
-export interface FileRouteTypes {
-  fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths: "/" | "/$extensionId/*"
-  fileRoutesByTo: FileRoutesByTo
-  to: "/" | "/$extensionId/*"
-  id: "__root__" | "/" | "/$extensionId/*"
-  fileRoutesById: FileRoutesById
-}
-
-export interface RootRouteChildren {
-  IndexRoute: typeof IndexRoute
-  ExtensionIdRoute: typeof ExtensionIdRoute
 }
 
 const rootRouteChildren: RootRouteChildren = {
   IndexRoute: IndexRoute,
-  ExtensionIdRoute: ExtensionIdRoute,
+  ExtensionIdSplatRoute: ExtensionIdSplatRoute,
 }
-
-export const routeTree = rootRoute._addFileChildren(rootRouteChildren)._addFileTypes<FileRouteTypes>()
-
-/* ROUTE_MANIFEST_START
-{
-  "routes": {
-    "__root__": {
-      "filePath": "__root.tsx",
-      "children": [
-        "/",
-        "/$extensionId/*"
-      ]
-    },
-    "/": {
-      "filePath": "index.tsx"
-    },
-    "/$extensionId/*": {
-      "filePath": "$extensionId.*.tsx"
-    }
-  }
-}
-ROUTE_MANIFEST_END */
+export const routeTree = rootRouteImport._addFileChildren(rootRouteChildren)._addFileTypes<FileRouteTypes>()