+++ apps/supernova/src/components/filters/PredefinedFilters.tsx
@@ -3,22 +3,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useState } from "react"
+import React from "react"
 
 import { Stack, TabNavigation, TabNavigationItem } from "@cloudoperators/juno-ui-components"
-import { useActivePredefinedFilter, useFilterActions, usePredefinedFilters } from "../StoreProvider"
+import { useNavigate } from "@tanstack/react-router"
+import { useActivePredefinedFilter, usePredefinedFilters } from "../StoreProvider"
 import SilenceScheduled from "../silences/SilenceScheduled"
 
 const PredefinedFilters = () => {
-  const { setActivePredefinedFilter } = useFilterActions()
+  const navigate = useNavigate()
   const predefinedFilters = usePredefinedFilters()
   const activePredefinedFilter = useActivePredefinedFilter()
 
-  const [selectedItem, setSelectedItem] = useState(activePredefinedFilter)
-
-  const handleTabSelect = (item: any) => {
-    setSelectedItem(item)
-    setActivePredefinedFilter(item)
+  const handleTabSelect = (item: React.ReactNode) => {
+    if (typeof item === "string") {
+      navigate({
+        to: "/alerts",
+        search: (prev) => ({
+          ...prev,
+          predefinedFilter: item,
+        }),
+      })
+    }
   }
 
   return (
@@ -33,8 +39,8 @@ const PredefinedFilters = () => {
   function renderPredefinedFilterTabs(): React.ReactNode {
     return (
       predefinedFilters &&
-      selectedItem && (
-        <TabNavigation activeItem={selectedItem} onActiveItemChange={handleTabSelect}>
+      activePredefinedFilter && (
+        <TabNavigation activeItem={activePredefinedFilter} onActiveItemChange={handleTabSelect}>
           {predefinedFilters?.map((filter: any) => (
             <TabNavigationItem key={filter.name} value={filter.name} label={filter.displayName} />
           ))}