+++ apps/supernova/src/lib/urlStateUtils.ts
@@ -1,0 +1,158 @@
+/*
+ * SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Juno contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+const ACTIVE_FILTERS = "f"
+const PAUSED_FILTERS = "pf"
+const ACTIVE_PREDEFINED_FILTER = "p"
+const DETAILS_FOR = "d"
+const SEARCH_TERM = "s"
+const ACTIVE_TAB = "t"
+const SILENCE_REG_EX = "r"
+const SILENCE_STATUS = "st"
+
+export const readLegacyUrlState = (state: any) => {
+  const activeFilters = state?.[ACTIVE_FILTERS]
+  const pausedFilters = state?.[PAUSED_FILTERS]
+  const searchTerm = state?.[SEARCH_TERM]
+  const predefinedFilter = !isNaN(state?.[ACTIVE_PREDEFINED_FILTER]) ? state?.[ACTIVE_PREDEFINED_FILTER] : undefined // for some reason old state return NaN if the value not present
+  const showDetailsFor = !isNaN(state?.[DETAILS_FOR]) ? state?.[DETAILS_FOR] : undefined // for some reason old state return NaN if the value not present
+  const activeTab = state?.[ACTIVE_TAB]
+  const silencesRegEx = state?.[SILENCE_REG_EX]
+  const silencesStatus = state?.[SILENCE_STATUS]
+
+  return {
+    activeFilters,
+    pausedFilters,
+    searchTerm,
+    predefinedFilter,
+    showDetailsFor,
+    activeTab,
+    silencesStatus,
+    silencesRegEx,
+  }
+}
+
+export const extractSearchStringFromHashFragment = (searchString: string) => {
+  /*
+   * In case of hashed routing Tanstack router returns URL search params of the entire URL rather than just from the hashed part.
+   * We'll have to extract the query part from the hash because otherwise in embedded mode the app will be taking search params from the shell app as well.
+   * Sanitize the search string by extracting the substring between the first '?' and the next '?' (if any), keeping the first '?'.
+   * https://github.com/TanStack/router/issues/4370
+   * http://localhost:3000/?preHashParam=prehashtest#/services?postHashParam1=test1?preHashParam=prehashtest
+   * searchString = "?postHashParam1=test1?preHashParam=prehashtest"
+   * searchStringFromHash = "?postHashParam1=test1"
+   */
+  const postHashParams = searchString.indexOf("?")
+  if (postHashParams === -1) return "" // If no query part is found, return an empty object
+  const preHashParams = searchString.indexOf("?", postHashParams + 1)
+  return searchString.slice(postHashParams, preHashParams === -1 ? undefined : preHashParams)
+}
+
+export const getFiltersForUrl = (prefix: string, filters: any) => {
+  return {
+    ...Object.entries(filters)
+      .map(([key, value]) => {
+        if (value === undefined || value === null) return {}
+        if (Array.isArray(value)) {
+          return { [`${prefix}${key}`]: value.join(",") }
+        }
+        return { [`${prefix}${key}`]: value }
+      })
+      .reduce((acc, curr) => ({ ...acc, ...curr }), {}),
+  }
+}
+
+export const convertAppStateToUrlState = (appState: any) => {
+  const activeFiltersForUrl = getFiltersForUrl("f_", appState.activeFilters || {})
+  const pausedFiltersForUrl = getFiltersForUrl("pf_", appState.pausedFilters || {})
+
+  return {
+    ...activeFiltersForUrl,
+    ...pausedFiltersForUrl,
+    predefinedFilter: appState.predefinedFilter,
+    searchTerm: appState.searchTerm,
+    showDetailsFor: appState.showDetailsFor,
+    silenceRegEx: appState.silenceRegEx,
+    silenceStatus: appState.silenceStatus,
+  }
+}
+
+export const getFiltersForApp = (prefix: string, urlState: Record<string, string | string[]>) => {
+  return Object.entries(urlState)
+    .filter(([key]) => key.startsWith(prefix))
+    .reduce((acc: Record<string, string[]>, [key, value]) => {
+      const filterKey = key.replace(prefix, "")
+      if (value === undefined || value === null) return acc
+      if (Array.isArray(value)) {
+        acc[filterKey] = value.map((v) => v.trim())
+      } else if (typeof value === "string") {
+        acc[filterKey] = [value.trim()]
+      } else {
+        return acc
+      }
+      return acc
+    }, {})
+}
+
+export const convertUrlStateToAppState = (urlState: any) => {
+  return {
+    activeFilters: getFiltersForApp("f_", urlState),
+    pausedFilters: getFiltersForApp("pf_", urlState),
+    predefinedFilter: urlState.predefinedFilter,
+    searchTerm: urlState.searchTerm,
+    showDetailsFor: urlState.showDetailsFor,
+    silencesRegEx: urlState.silencesRegEx,
+    silencesStatus: urlState.silencesStatus,
+  }
+}
+
+// removes a specific filter from the filters object and returns a new object
+export const removeFilter = (filters: any, filterKey: string, filterValue: string) => {
+  // if filter has more than one value, remove the specific value
+  if (Array.isArray(filters?.[filterKey])) {
+    const updatedFilters: any = {
+      ...filters,
+      [filterKey]: filters?.[filterKey].filter((value: string) => value !== filterValue),
+    }
+    if (updatedFilters?.[filterKey]?.length === 0) {
+      delete updatedFilters[filterKey]
+    }
+    return updatedFilters
+  }
+
+  // if filter is a single value, remove the filter key
+  const updatedFilters = { ...filters }
+  if (updatedFilters?.[filterKey] === filterValue) delete updatedFilters[filterKey]
+
+  return updatedFilters
+}
+
+// adds a specific filter to the filters object and returns a new object
+export const addFilter = (filters: any, filterKey: string, filterValue: string) => {
+  // if the filter already exists, add the value to the existing array
+  if (filters?.[filterKey] && Array.isArray(filters[filterKey])) {
+    // Create a Set from existing values to remove duplicates
+    const filterSet = new Set(filters[filterKey])
+    // Add the new value
+    filterSet.add(filterValue)
+    // Convert back to array
+    return {
+      ...filters,
+      [filterKey]: Array.from(filterSet),
+    }
+  }
+  // If the filter is a single value, convert it to an array
+  if (filters?.[filterKey] && typeof filters[filterKey] === "string") {
+    return {
+      ...filters,
+      [filterKey]: [filters[filterKey], filterValue],
+    }
+  }
+  // If the filter does not exist, create a new key with the value
+  return {
+    ...filters,
+    [filterKey]: filterValue,
+  }
+}