+++ apps/supernova/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @cloudoperators/juno-app-supernova
 
+## 0.19.1
+
+### Patch Changes
+
+- 9e9669f: Combine initial filters from appProps with fitlers from the url state.
+- Updated dependencies [1020d89]
+  - @cloudoperators/juno-ui-components@5.3.0
+  - @cloudoperators/juno-messages-provider@0.2.14
+
+## 0.19.0
+
+### Minor Changes
+
+- 8a44efa: Supernova old url state is now converted to new URL state and for other apps tanstack router dependencies are updated.
+
+### Patch Changes
+
+- Updated dependencies [b7fe310]
+  - @cloudoperators/juno-ui-components@5.2.0
+  - @cloudoperators/juno-messages-provider@0.2.13
+
 ## 0.18.0
 
 ### Minor Changes